requirements.txt
# Core dependencies
python-binance==1.0.19
ccxt==4.1.22
asyncpg==0.29.0
aiohttp==3.9.1
python-dotenv==1.0.0
pyyaml==6.0.1

# Bybit
pybit==5.6.2

# Database
sqlalchemy==2.0.23
alembic==1.13.0

# Monitoring
prometheus-client==0.19.0
psutil==5.9.6

# Utils
numpy
pandas==2.0.3
scipy==1.11.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
ipython==8.18.1